[DEV-1900] Refactoring
