+++ src/subdomains/core/buy-crypto/process/dto/update-buy-crypto.dto.ts
@@ -146,4 +146,9 @@ export class UpdateBuyCryptoDto {
   @IsOptional()
   @IsNumber()
   allowedTotalFeePercent: number;
+
+  @IsOptional()
+  @IsDate()
+  @Type(() => Date)
+  payoutConfirmationDate: Date;
 }
