+++ src/subdomains/core/buy-crypto/process/services/buy-crypto-batch.service.ts
@@ -18,14 +18,12 @@ import { PriceRequest, PriceResult } from 'src/subdomains/supporting/pricing/dom
 import { PricingService } from 'src/subdomains/supporting/pricing/services/pricing.service';
 import { In, IsNull, Not } from 'typeorm';
 import { BuyCryptoBatch, BuyCryptoBatchStatus } from '../entities/buy-crypto-batch.entity';
-import { BuyCryptoFee } from '../entities/buy-crypto-fees.entity';
 import { BuyCrypto, BuyCryptoStatus } from '../entities/buy-crypto.entity';
 import { MissingBuyCryptoLiquidityException } from '../exceptions/abort-batch-creation.exception';
 import { BuyCryptoBatchRepository } from '../repositories/buy-crypto-batch.repository';
 import { BuyCryptoRepository } from '../repositories/buy-crypto.repository';
 import { BuyCryptoNotificationService } from './buy-crypto-notification.service';
 import { BuyCryptoPricingService } from './buy-crypto-pricing.service';
-import { BuyCryptoWebhookService } from './buy-crypto-webhook.service';
 
 @Injectable()
 export class BuyCryptoBatchService {
@@ -41,52 +39,7 @@ export class BuyCryptoBatchService {
     private readonly payoutService: PayoutService,
     private readonly buyCryptoNotificationService: BuyCryptoNotificationService,
     private readonly liquidityService: LiquidityManagementService,
-    private readonly buyCryptoWebhookService: BuyCryptoWebhookService,
   ) {}
-
-  async prepareTransactions(): Promise<void> {
-    try {
-      const txInput = await this.buyCryptoRepo.find({
-        where: {
-          inputReferenceAmountMinusFee: Not(IsNull()),
-          outputReferenceAsset: IsNull(),
-          outputAsset: IsNull(),
-          batch: IsNull(),
-        },
-        relations: [
-          'bankTx',
-          'buy',
-          'buy.user',
-          'buy.user.wallet',
-          'buy.asset',
-          'batch',
-          'cryptoRoute',
-          'cryptoRoute.user',
-          'cryptoRoute.user.wallet',
-          'cryptoRoute.asset',
-          'cryptoInput',
-        ],
-      });
-
-      if (txInput.length === 0) return;
-
-      this.logger.verbose(
-        `Buy-crypto transaction input. Processing ${txInput.length} transaction(s). Transaction ID(s): ${txInput.map(
-          (t) => t.id,
-        )}`,
-      );
-
-      const txWithAssets = await this.defineAssetPair(txInput);
-      const txWithFeeConstraints = this.setFeeConstraints(txWithAssets);
-
-      for (const tx of txWithFeeConstraints) {
-        await this.buyCryptoRepo.save(tx);
-        await this.buyCryptoWebhookService.triggerWebhook(tx);
-      }
-    } catch (e) {
-      this.logger.error('Error during buy-crypto preparation:', e);
-    }
-  }
 
   async batchAndOptimizeTransactions(): Promise<void> {
     try {
@@ -137,45 +90,6 @@ export class BuyCryptoBatchService {
     } catch (e) {
       this.logger.error('Error during buy-crypto batching:', e);
     }
-  }
-
-  private async defineAssetPair(transactions: BuyCrypto[]): Promise<BuyCrypto[]> {
-    for (const tx of transactions) {
-      try {
-        const outputReferenceAssetToFetch = tx.defineAssetExchangePair();
-
-        if (outputReferenceAssetToFetch) {
-          const { outputReferenceAssetName, type } = outputReferenceAssetToFetch;
-
-          const outputReferenceAsset = await this.assetService.getAssetByQuery({
-            dexName: outputReferenceAssetName,
-            blockchain: tx.outputAsset.blockchain,
-            type,
-          });
-
-          if (!outputReferenceAsset) {
-            throw new Error(
-              `Asset with name ${outputReferenceAssetName}, type: ${type}, blockchain: ${tx.outputAsset.blockchain} not found by asset service.`,
-            );
-          }
-
-          tx.setOutputReferenceAsset(outputReferenceAsset);
-        }
-      } catch (e) {
-        this.logger.error('Error while defining buy-crypto asset pair:', e);
-      }
-    }
-
-    return transactions.filter((tx) => tx.outputReferenceAsset && tx.outputAsset);
-  }
-
-  private setFeeConstraints(transactions: BuyCrypto[]): BuyCrypto[] {
-    for (const tx of transactions) {
-      const fee = BuyCryptoFee.create(tx);
-      tx.setFeeConstraints(fee);
-    }
-
-    return transactions;
   }
 
   private async getReferencePrices(txWithAssets: BuyCrypto[]): Promise<Price[]> {
